Move data portion of `SavedQuery` class to dbt/artifacts
