+++ .changes/unreleased/Under the Hood-20240126-164038.yaml
@@ -1,0 +1,6 @@
+kind: Under the Hood
+body: Move `SavedQuery` data definition to `dbt/artifacts`
+time: 2024-01-26T16:40:38.790993-08:00
+custom:
+  Author: QMalcolm
+  Issue: "9386"
